requirements.txt
python-binance==1.0.19
requests==2.31.0
python-dotenv==1.0.0
colorama==0.4.6
tabulate==0.9.0
pandas>=2.0.0
numpy>=1.24.0
